timeline bugfix

+++ app/controllers/QuestsController.php
@@ -19,7 +19,9 @@ class QuestsController extends \BaseController {
 				$quest = Quest::where("user_id", "=", $user->id)->where("id", "=", $quest_id)->first();
 				if($quest) {
 					$timeline = Timeline::where("user_id", "=", $user->id)->where('quest_id', '=', $quest_id)->first();
-					$timeline->delete();
+					if($timeline) {
+						$timeline->delete();
+					}
 					$quest->delete();
 					$time = date("U");
 					if(($quest->createDate + 86400000) > ($time*1000)) {
